+++ src/test/java/model/card/RandomCardTest.java
@@ -7,14 +7,14 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-class CardDispenserTest {
+class RandomCardTest {
 
-    RandomCardPicker cardDispenser = RandomCardPicker.getRandomCardPicker();
+    RandomCard randomCard = RandomCard.getRandomCard();
 
     @DisplayName("랜덤으로 숫자와 모양을 뽑아 카드 1장을 지급한다")
     @Test
     void testDispenseCard() {
-        Card card = cardDispenser.pickCard();
+        Card card = randomCard.pickCard();
         assertThat(card.getNumber()).isNotNull();
         assertThat(card.getShape()).isNotNull();
     }
@@ -23,7 +23,7 @@ class CardDispenserTest {
     @ParameterizedTest
     @ValueSource(ints = {0, 1, 5, 10, 29, 60})
     void testDispenseCards(int count) {
-        Cards cards = cardDispenser.pickCards(count);
+        Cards cards = randomCard.pickCards(count);
         assertThat(cards.getElements()).hasSize(count);
         cards.getElements().forEach(card -> {
             assertThat(card.getNumber()).isNotNull();
